+++ lib/src/models/config_models/emoji_message_configuration.dart
@@ -25,6 +25,7 @@ class EmojiMessageConfiguration {
   const EmojiMessageConfiguration({
     this.padding,
     this.textStyle,
+    this.maxOutSideBubbleEmojis,
   });
 
   /// Used for giving padding to emoji messages.
@@ -32,4 +33,8 @@ class EmojiMessageConfiguration {
 
   /// Used for giving text style to emoji messages.
   final TextStyle? textStyle;
+
+  /// Maximum number of emojis to show outside the bubble
+  /// for emoji-only messages.
+  final int? maxOutSideBubbleEmojis;
 }